+++ architecture_tool_django/graphdefs/urls.py
@@ -4,7 +4,6 @@ from . import views
 
 app_name = "graphs"
 urlpatterns = [
-    path("graphcount/", views.graphdef_count, name="graphdef.count"),
     path("graphs/", views.GraphListView.as_view(), name="graph.list"),
     path("graphs/create/", views.GraphCreateView.as_view(), name="graph.create"),
     path(
@@ -22,5 +21,4 @@ urlpatterns = [
         views.GraphDeleteView.as_view(),
         name="graph.delete",
     ),
-    path("graphapi/<str:pk>/", views.get_graph, name="graphapi"),
 ]