+++ architecture_tool_django/listdefs/urls.py
@@ -4,7 +4,6 @@ from . import views
 
 app_name = "lists"
 urlpatterns = [
-    path("listcount/", views.listdef_count, name="listdef.count"),
     path("lists/", views.ListdefListView.as_view(), name="listdef.list"),
     path("lists/create/", views.ListdefCreateView.as_view(), name="listdef.create"),
     path(
@@ -22,5 +21,4 @@ urlpatterns = [
         views.ListdefDeleteView.as_view(),
         name="listdef.delete",
     ),
-    path("listapi/<str:pk>/", views.get_listdef, name="listdefapi"),
 ]