+++ config/api_router.py
@@ -1,6 +1,8 @@
 from django.conf import settings
 from rest_framework.routers import DefaultRouter, SimpleRouter
 
+from architecture_tool_django.graphdefs.api.views import GraphViewSet
+from architecture_tool_django.listdefs.api.views import ListViewSet
 from architecture_tool_django.modeling.api.views import (
     EdgetypeViewSet,
     NodetypeViewSet,
@@ -20,6 +22,8 @@ router.register("schemas", SchemaViewSet)
 router.register("nodetypes", NodetypeViewSet)
 router.register("edgetypes", EdgetypeViewSet)
 router.register("nodes", NodeViewSet)
+router.register("lists", ListViewSet)
+router.register("graphs", GraphViewSet)
 
 app_name = "api"
 urlpatterns = router.urls